python/requirements.txt
pycapnp>=1.0.0
torch>=2.0.0
numpy>=1.24.0
click>=8.1.0
pyyaml>=6.0

